Function actix_threadpool::run

pub fn run<F, I, E>(f: F) -> CpuFuture<I, E> where
    F: FnOnce() -> Result<I, E> + Send + 'static,
    I: Send + 'static,
    E: Send + Debug + 'static, 

Execute blocking function on a thread pool, returns future that resolves to result of the function execution.